+++ yarn-project/end-to-end/src/e2e_p2p/reqresp.test.ts
@@ -1,9 +1,6 @@
-import type { Archiver } from '@aztec/archiver';
 import type { AztecNodeService } from '@aztec/aztec-node';
 import { sleep } from '@aztec/aztec.js';
 import { RollupAbi } from '@aztec/l1-artifacts';
-import type { SequencerClient } from '@aztec/sequencer-client';
-import { BlockAttestation, ConsensusPayload } from '@aztec/stdlib/p2p';
 
 import { jest } from '@jest/globals';
 import fs from 'fs';
@@ -126,25 +123,6 @@ describe('e2e_p2p_reqresp_tx', () => {
       ),
     );
     t.logger.info('All transactions mined');
-
-    // Gather signers from attestations downloaded from L1
-    const blockNumber = await contexts[0].txs[0].getReceipt().then(r => r.blockNumber!);
-    const dataStore = ((nodes[0] as AztecNodeService).getBlockSource() as Archiver).dataStore;
-    const [block] = await dataStore.getBlocks(blockNumber, blockNumber);
-    const attestations = block.signatures.map(
-      sig => new BlockAttestation(ConsensusPayload.fromBlock(block.block), sig),
-    );
-    const signers = await Promise.all(attestations.map(att => att.getSender().then(s => s.toString())));
-    t.logger.info(`Attestation signers`, { signers });
-
-    // Check that the signers found are part of the proposer nodes to ensure the archiver fetched them right
-    const validatorAddresses = nodes.map(node =>
-      ((node as AztecNodeService).getSequencer() as SequencerClient).validatorAddress?.toString(),
-    );
-    t.logger.info(`Validator addresses`, { addresses: validatorAddresses });
-    for (const signer of signers) {
-      expect(validatorAddresses).toContain(signer);
-    }
   });
 
   /**
